[PF-819] Flesh out tests for cloning controlled GCS bucket (#930)
* [PF-819] Flesh out tests for cloning controlled GCS bucket

[PF-819] Flesh out tests for cloning controlled GCS bucket

[PF-819] Flesh out tests for cloning controlled BQ dataset

[PF-819] Flesh out tests for cloning controlled BQ dataset

same vs different workspace

move 1 test

move to controller tests

duplicate dataset name

copy_nothing, copy_definition

fix tests

pass in GoogleCredential

delete cloneBqDataset_requesterNoReadPermissionOnSourceWorkspace_throws403

add retry

take clone out of ControlledBigQueryDatasetLifecycle

access control tests

location

minor fixes

spotless

spotbugs

go back to CRL 1.2.2

add back destinationWorkspaceId and fix test not using it

[PF-819] Flesh out tests for cloning controlled GCS bucket

set and assert GCP attributes

deflake

uncomment

fix notebook test

* fix tests

+++ integration/src/main/java/scripts/testscripts/ControlledBigQueryDatasetLifecycle.java
@@ -195,7 +195,7 @@ public class ControlledBigQueryDatasetLifecycle extends GcpWorkspaceCloneTestScr
 
     // Workspace owner can update the dataset resource through WSM
     String resourceDescription = "a description for WSM";
-    Integer defaultTableLifetimeSec = 5400;
+    Long defaultTableLifetimeSec = 5400L;
     var updateDatasetRequest =
         new UpdateControlledGcpBigQueryDatasetRequestBody()
             .description(resourceDescription)
